refactor: move some leftover APIs to EpochManagerAdapter
